10_MultiAgent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic>=2.0
mysql-connector-python==8.3.0
openai>=1.30.0
tavily-python>=0.3.0
chromadb>=0.5.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
flashrank>=0.2.0
kiwipiepy>=0.17.0
python-dotenv==1.0.1
requests>=2.31.0
